[nan-1762] fix microsoft
